clean up
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@23702 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 05ec3ad2ed583235789c5572924aad4d72baa905

+++ sandbox/rf_detector/include/librf/discrete_dist.h
@@ -36,7 +36,7 @@ namespace librf {
 		{
 			dist.resize(counter_.size());
 			assert(dist.size() == counter_.size());
-			for (int i = 0; i < dist.size(); i++)
+			for (int i = 0; i < (int)(dist.size()); i++)
 			{
 				dist[i] = counter_[i] / (float)sum_;
 			}
